Sending emails(just simulations), modifying MessageDialogView

+++ Modules/Outlook.Modules.Mail/ViewModels/MailListViewModel.cs
@@ -7,6 +7,7 @@ using Prism.Commands;
 using Prism.Regions;
 using Prism.Services.Dialogs;
 using System.Collections.ObjectModel;
+using System.Windows;
 using MailMessage = Outlook.Business.MailMessage;
 
 namespace Outlook.Modules.Mail.ViewModels
@@ -19,6 +20,7 @@ namespace Outlook.Modules.Mail.ViewModels
         // Service to work with dialogs
         private readonly IRegionDialogService _dialogService;
 
+        private string _currentFolder = FolderParameters.Inbox;
 
         private string _title = "Default";
         public string Title
@@ -47,14 +49,18 @@ namespace Outlook.Modules.Mail.ViewModels
         /// Command to show UI dialog for writing, sending emails
         /// </summary>
 
-        private DelegateCommand _messageCommand;
-        public DelegateCommand MessageCommand =>
-            _messageCommand ?? (_messageCommand = new DelegateCommand(ExecuteMessageCommand));
-
-
         private DelegateCommand _newMessageCommand;
         public DelegateCommand NewMessageCommand =>
             _newMessageCommand ?? (_newMessageCommand = new DelegateCommand(ExecuteNewMessageCommand));
+
+
+        private DelegateCommand _replyMessageCommand;
+        public DelegateCommand ReplyMessageCommand =>
+            _replyMessageCommand ?? (_replyMessageCommand = new DelegateCommand(ExecuteReplyMessageCommand));
+
+        private void ExecuteReplyMessageCommand()
+        {
+        }
 
         private void ExecuteNewMessageCommand()
         {
@@ -66,12 +72,9 @@ namespace Outlook.Modules.Mail.ViewModels
                 parameters,
                 dialogResult =>
                 {
-
+                    SetCorretMessageFromDialog(dialogResult);
                 });
         }
-
-
-        public event Action<IDialogResult>? RequestClose;
 
         /// <summary>
         /// Command to delete selected mail message
@@ -84,6 +87,12 @@ namespace Outlook.Modules.Mail.ViewModels
             if(SelectedMailMessage == null)
                 return;
 
+            var messageBoxResult = MessageBox.Show("Are you sure you want to delete this message?", 
+                                                        "Delete message", MessageBoxButton.YesNo);
+
+            if (messageBoxResult == MessageBoxResult.No)
+                return;
+
             // Remove from the database
             _mailService.DeleteMessageById(SelectedMailMessage.Id);
 
@@ -92,25 +101,7 @@ namespace Outlook.Modules.Mail.ViewModels
         }
 
 
-        // show UI dialog for writing, sending emails
-        private void ExecuteMessageCommand()
-        {
-            if (SelectedMailMessage == null)
-            {
-                return;
-            }
-
-            var parameters = new DialogParameters();
-            parameters.Add(FolderParameters.MailMessageKey, SelectedMailMessage.Id);
-            // show dialog 
-            _dialogService.ShowRegionDialog(RegionNames.ContentRegion, 
-                nameof(MessageDialogView), 
-                parameters,
-                dialogResult =>
-                {
-
-                });
-        }
+        public event Action<IDialogResult>? RequestClose;
 
         public MailListViewModel(IMailService mailService,
             IRegionDialogService dialogService)
@@ -128,23 +119,43 @@ namespace Outlook.Modules.Mail.ViewModels
         public override void OnNavigatedTo(NavigationContext navigationContext)
         {
             // according the key we can get value
-            var folder = navigationContext.Parameters.GetValue<string>(FolderParameters.FolderKey);
+            _currentFolder = navigationContext.Parameters.GetValue<string>(FolderParameters.FolderKey);
 
+            //
+            LoadMessages(_currentFolder);
+
+            // set the first mail message as selected
+            SelectedMailMessage = MailMessages.FirstOrDefault();
+        }
+
+
+        void LoadMessages(string folder)
+        {
             switch (folder)
             {
                 case FolderParameters.Inbox:
                     MailMessages = new ObservableCollection<MailMessage>(_mailService.GetMailMessages());
                     break;
                 case FolderParameters.Sent:
-                    MailMessages = new ObservableCollection<MailMessage>(_mailService.SentMailMailMessages());
+                    MailMessages = new ObservableCollection<MailMessage>(_mailService.GetSentMailMailMessages());
                     break;
                 case FolderParameters.Deleted:
                     MailMessages = new ObservableCollection<MailMessage>(_mailService.GetDeletedMessages());
                     break;
             }
+        }
 
-            // set the first mail message as selected
-            SelectedMailMessage = MailMessages.FirstOrDefault();
+        /// <summary>
+        /// Set correct message from dialog
+        /// </summary>
+        /// <param name="dialogResult"></param>
+        void SetCorretMessageFromDialog(IDialogResult dialogResult)
+        {
+            if (dialogResult.Parameters.TryGetValue<MailMessage>(FolderParameters.MessageSent, out var message))
+            {
+                // add to sent folder
+                MailMessages.Add(message);
+            }
         }
     }
 }
